Moved UnclosableOutputStream class into CommandScope and made it private Do not create if output stream is null
DAT-10419
