+++ liquibase-core/src/main/java/liquibase/command/CommandScope.java
@@ -4,9 +4,10 @@ import liquibase.Scope;
 import liquibase.configuration.*;
 import liquibase.exception.CommandExecutionException;
 import liquibase.exception.CommandValidationException;
-import liquibase.io.UnclosableOutputStream;
 import liquibase.util.StringUtil;
 
+import java.io.FilterOutputStream;
+import java.io.IOException;
 import java.io.OutputStream;
 import java.util.*;
 
@@ -129,11 +130,15 @@ public class CommandScope {
      */
     public CommandScope setOutput(OutputStream outputStream) {
         /*
-        This is an UncloseableOutputStream because we do not want individual command steps to inadvertently (or
+        This is an UnclosableOutputStream because we do not want individual command steps to inadvertently (or
         intentionally) close the System.out OutputStream. Closing System.out renders it unusable for other command
-        steps which expect it to still be open.
+        steps which expect it to still be open.  If the passed OutputStream is null then we do not create it.
          */
-        this.outputStream = new UnclosableOutputStream(outputStream);
+        if (outputStream != null) {
+            this.outputStream = new UnclosableOutputStream(outputStream);
+        } else {
+            this.outputStream = null;
+        }
 
         return this;
     }
@@ -201,6 +206,29 @@ public class CommandScope {
                 .setValueHandler(argument.getValueConverter())
                 .setValueObfuscator(argument.getValueObfuscator())
                 .buildTemporary();
+    }
+
+    /**
+     * This class is a wrapper around OutputStreams, and makes them impossible for callers to close.
+     */
+    private static class UnclosableOutputStream extends FilterOutputStream {
+        public UnclosableOutputStream(OutputStream out) {
+            super(out);
+        }
+
+        @Override
+        public void write(byte[] b, int off, int len) throws IOException {
+            out.write(b, off, len);
+        }
+
+        /**
+         * This method does not actually close the underlying stream, but rather only flushes it. Callers should not be
+         * closing the stream they are given.
+         */
+        @Override
+        public void close() throws IOException {
+            out.flush();
+        }
     }
 
     private class CommandScopeValueProvider extends AbstractMapConfigurationValueProvider {
